Complete unit and integratiob tests for home page
Test includes smoke and snapshot test of initial Home page
component as well as expectation tests for flights data
on the page. Have not set up state management for cases where
API calls fail or the data set is empty, but would be the next
task if the data fetching layer grows.